mcq_generator/requirements.txt
Flask
openai
python-docx
PyPDF2